fix(automations): clear lint-ratchet violations on movement triggers
CI's lint:ci ratchet failed on the new became-mobile / left-home code:

- no-explicit-any: replace `catch (e: any)` with `catch (e)` +
  `e instanceof Error ? e.message : String(e)` in leftHomeFromHistory,
  automationEngineService, automationRoutes and AutomationsPage; type
  `rowToRecord(row: Record<string, unknown>)` in the home-anchors repo.
- prefer-const: `let home` / `let inner` in checkLeftHome are never
  reassigned → const.
- react-refresh/only-export-components: move non-component exports out of
  component files into pure sibling modules —
  `leftHome*`/`numOrUndef`/`EventState`/`FactState` → automationTesterHelpers.ts,
  `labelOf` → nodeLabel.ts, `NODE_TOKENS`/`SUBJECT_NODE_TRIGGER_TYPES` →
  substitutionNodeTokens.ts. Importers (tokenHints, tests) updated.

No behavior change. lint:ci OK, tsc clean, affected suites green.
