+++ src/components/automations/SubstitutionsHelp.tsx
@@ -7,6 +7,7 @@
  */
 import { UiIcon } from '../icons';
 import { HOP_COUNT_EMOJIS, HOP_EMOJI_MAX, MQTT_SOURCE_EMOJI } from '../../utils/hopEmoji';
+import { NODE_TOKENS, SUBJECT_NODE_TRIGGER_TYPES } from './substitutionNodeTokens';
 
 // All `{{ trigger.* }}` tokens, by trigger type. `sourceId`/`timestamp` are added to every group.
 export const TRIGGER_TOKENS: Record<string, Array<[string, string]>> = {
@@ -45,38 +46,6 @@ export const TRIGGER_TOKENS: Record<string, Array<[string, string]>> = {
   'trigger.leftHome': [['nodeNum', 'Node number'], ['latitude', 'Node latitude'], ['longitude', 'Node longitude'], ['homeLat', 'Home latitude'], ['homeLon', 'Home longitude'], ['distanceMeters', 'Distance from home (m)'], ['thresholdMeters', 'Configured threshold (m)']],
   'trigger.schedule': [],
 };
-
-/**
- * Hydrated subject-node tokens (`{{ node.longName }}`, …). Available on every
- * trigger that has a subject node — same paths conditions use in fieldselect.
- * Not under `trigger.*`: they hydrate from the DB node row at evaluation time.
- */
-export const NODE_TOKENS: Array<[string, string]> = [
-  ['longName', 'Node long name'],
-  ['shortName', 'Node short name'],
-  ['nodeId', 'Node id (!hex)'],
-  ['roleName', 'Role name (e.g. ROUTER)'],
-  ['batteryLevel', 'Battery level (%)'],
-  ['voltage', 'Voltage'],
-  ['hopsAway', 'Hops away'],
-  ['latitude', 'Latitude'],
-  ['longitude', 'Longitude'],
-  ['altitude', 'Altitude'],
-  ['mobile', 'Mobile flag (1 = mobile, 0 = stationary)'],
-  ['ageMinutes', 'Minutes since last heard'],
-  ['completeness', 'Node info completeness (complete / incomplete / unknown)'],
-];
-
-/** Triggers whose event carries a subject node (so `{{ node.* }}` can hydrate). */
-export const SUBJECT_NODE_TRIGGER_TYPES = [
-  'trigger.message',
-  'trigger.nodeDiscovered',
-  'trigger.nodeUpdated',
-  'trigger.telemetry',
-  'trigger.geofence',
-  'trigger.becameMobile',
-  'trigger.leftHome',
-] as const;
 
 export const UNIVERSAL_TOKENS: Array<[string, string]> = [['sourceId', 'The source the event came from'], ['timestamp', 'Event time (rendered as a local date/time)']];
 const TRIGGER_LABEL: Record<string, string> = {
